+++ migrations/20170115200001_initial_schema.sql
@@ -37,8 +37,7 @@ CREATE TABLE IF NOT EXISTS users (
     created_at     BIGINT        CHECK (created_at > 0) NOT NULL,
     updated_at     BIGINT        CHECK (updated_at > 0) NOT NULL,
     verified_at    BIGINT        CHECK (verified_at >= 0) DEFAULT 0 NOT NULL,
-    disabled_at    BIGINT        CHECK (disabled_at >= 0) DEFAULT 0 NOT NULL,
-    last_online_at BIGINT        CHECK (last_online_at >= 0) DEFAULT 0 NOT NULL
+    disabled_at    BIGINT        CHECK (disabled_at >= 0) DEFAULT 0 NOT NULL
 );
 
 -- This table should be replaced with an array column in the users table