+++ server/pipeline_user.go
@@ -42,7 +42,7 @@ func (p *pipeline) usersFetch(logger *zap.Logger, session session, envelope *Env
 		}
 	}
 
-	users, err := UsersFetchIdsHandles(logger, p.db, userIds, handles)
+	users, err := UsersFetchIdsHandles(logger, p.db, p.tracker, userIds, handles)
 	if err != nil {
 		logger.Warn("Could not retrieve users", zap.Error(err))
 		session.Send(ErrorMessageRuntimeException(envelope.CollationId, "Could not retrieve users"), true)